TS_1/requirements.txt
pytest

selenium

pytest-selenium

pytest-allure

page-objects